Refactor demos, add demo dropdown and conditionally render param ui based on plotType
